requirements.txt
Django~=4.2.1
mysql~=0.0.3
mysqlclient~=2.2.0
pandas~=2.0.2
Pillow~=9.5.0
pycryptodome~=3.17
openpyxl~=3.1.2
uWSGI~=2.0.21